modified: refactor the check permission function
