+++ src/pages/AdminUserPage.jsx
@@ -1,11 +1,12 @@
+// package
 import styled from 'styled-components'
 import { useState, useEffect } from 'react'
-import { useNavigate } from "react-router-dom";
+// component and style
 import AdminUserCard from '../components/admin/AdminUserCard'
 import * as style from '../components/common/common.styled'
 // api
 import { adminGetUsers } from '../api/admin'
-import { checkAdminPermission } from '../api/Permission'
+import { useAdminAuthValitate } from '../utils/authValidate';
 
 const Container = styled.div`
   width: 83%;
@@ -37,22 +38,8 @@ const CardContainer = styled.div`
 
 export default function AdminUserPage(){
     const [users, setUsers] = useState([])
-    const navigate = useNavigate()
     // 驗證 token
-    useEffect(() => {
-      const checkTokenIsValid = async () => {
-        const authToken = localStorage.getItem('AdminToken');
-        if (!authToken) {
-          navigate('/admin_login');
-        }
-        const result = await checkAdminPermission(authToken);
-        if (!result) {
-          navigate('/admin_login');
-        }
-      };
-
-      checkTokenIsValid();
-    }, [navigate])
+    useAdminAuthValitate('/admin_login')
     // 取得所有用戶
     useEffect(() => {
     const fetchUsers = async () => {
