+++ src/pages/LoginPage.jsx
@@ -1,3 +1,7 @@
+// package
+import { useState, useEffect, useContext } from "react";
+import { Link, useNavigate } from "react-router-dom";
+// component and style
 import {
   AuthContainer,
   AuthInputContainer,
@@ -5,21 +9,16 @@ import {
   AuthLinkText,
   AuthLinkWrapper,
   AuthTittle,
-  AuthSpan,
-  InputLength,
-  WarnMsg,
+  AuthSpan
 } from "../components/common/auth.styled";
 import { ReactComponent as Logo } from "../assets/logo.svg";
 import AuthInput from "../components/AuthInput";
-import { useState, useEffect, useContext } from "react";
-import { Link, useNavigate } from "react-router-dom";
-// api
-import { checkPermission } from "../api/Permission";
+import { Toast } from "../components/common/common.styled";
+// api and function
 import { login } from "../api/auth";
 import { getUser } from "../api/user";
-import clsx from "clsx";
 import { UserContext } from "../context/UserContext";
-import { Toast } from "../components/common/common.styled";
+import { useAuthValitate } from "../utils/authValidate";
 
 
 
@@ -27,8 +26,9 @@ const LoginPage = () => {
   const [account, setAccount] = useState("");
   const [password, setPassword] = useState("");
   const { setUserData } = useContext(UserContext);
-
   const navigate = useNavigate();
+  // 驗證 token
+  useAuthValitate('/main')
 
   // error control
   const [accountError, setAccountError] = useState("")
@@ -83,31 +83,13 @@ const LoginPage = () => {
     }
   };
 
-  // 驗證 token
-  useEffect(() => {
-    const checkTokenIsValid = async () => {
-      const authToken = localStorage.getItem("UserToken");
-      if (!authToken) {
-        return;
-      }
-      const result = await checkPermission(authToken);
-      if (result) {
-        navigate("/main");
-      }
-    };
-
-    checkTokenIsValid();
-  }, [navigate]);
-
   return (
     <AuthContainer>
       <div>
         <Logo />
       </div>
       <AuthTittle>登入Alphitter</AuthTittle>
-      <AuthInputContainer
-        className={clsx("", { redLine: account.length > 30 })}
-      >
+      <AuthInputContainer>
         <AuthInput
           label={"帳號"}
           maxLength={30}
@@ -120,9 +102,7 @@ const LoginPage = () => {
         />
       </AuthInputContainer>
 
-      <AuthInputContainer
-        className={clsx("", { redLine: password.length > 20 })}
-      >
+      <AuthInputContainer>
         <AuthInput
           label={"密碼"}
           type={'password'}
